* Added clean UI allowing for on the fly modification of renderer settings. This has introduced dependencies on css/slider.css and js/slider.js. A very basic reset style sheet has also been added as reset.css.
* All dependencies downloaded and commited along with the project files
for quick local loading and future breakage prevention. This includes
jquery (1.6.2) and compatible jquery.tools for tooltip support. The
tooltips also requires a background .png placed in /images.

* Numerous performance optimizations in the raytracer. Most notably
removal of this/self references from the core loops and introduction
of precalc LUTs where appropriate. All calculations pushed to
outermost possible scope and local variable caching introduced as
and when it was found to constitute an improvement. Methods with a
single callee inlined for added performance. The recursion level has
been limited to four, as this was found to be the highest visually
beneficial detail level at our current resolution.